Use interface for mock
